test(layout,render): migrate inline tests to integration tests
- Layout module: 35 tests migrated to tests/layout_tests.rs
  - Rect and merge_rects tests (public API)
  - Responsive layout tests (public API)
  - LayoutEngine public API tests
  - Tests accessing private internals kept inline (node, tree, flex, grid, etc.)

- Render module: 70 tests migrated to tests/render_tests.rs
  - Cell and Modifier tests
  - Buffer tests
  - Diff tests
  - Batch tests (except private field access)
  - Image protocol tests (except private methods)
  - Backend capabilities tests
  - Tests accessing private internals kept inline (terminal, crossterm)

Part of #197
